+++ src/cli/mod.rs
@@ -2,6 +2,7 @@ mod action;
 mod args;
 mod help;
 mod instance;
+mod port;
 mod printer;
 mod run;